requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic-settings==2.8.1
pandas==2.1.4